fix3

+++ src/modules.py
@@ -672,7 +672,7 @@ class NormalizationModule(OperationModule):
       ?:                    tensor, same shape as x
     """
     casted_x = tf.cast(x, dtype=self.dtype)
-    rescaled_x = (casted_x / 255) * (self.inp_max - self.inp_min) - abs(self.inp_min)
+    rescaled_x = (casted_x / 255) * (self.inp_max - self.inp_min) -  tf.cast(abs(self.inp_min), dtype=self.dtype)
     return rescaled_x
